Refactor project :tada: Controlling the lights is now managed by a single class, passed down through the dependency tree.
