+++ src/main/kotlin/tech/johari/gamelights/FlashThread.kt
@@ -1,22 +1,23 @@
 package tech.johari.gamelights
 
-import com.github.kittinunf.fuel.Fuel
-import tech.johari.gamelights.Main
+import tech.johari.gamelights.lights.LightManager
+import tech.johari.gamelights.lights.LightSignal
 
-class FlashThread(val color: LightSignal, val speed: Long) : Thread() {
+class FlashThread(val lights: LightManager, val color: LightSignal, val speed: Long) : Thread() {
+
     var state = true
-    val startTime = System.currentTimeMillis()
 
     override fun run() {
         if(!state) {
             // Turn it on
-            Fuel.get("http://${Main.PI_LOCAL_IP}:${Main.PI_API_PORT}/ha-api?cmd=1&scope=${color.name.toLowerCase()}").response { request, response, result -> }
+            lights.update(color, true)
         } else {
             // Turn it off
-            Fuel.get("http://${Main.PI_LOCAL_IP}:${Main.PI_API_PORT}/ha-api?cmd=0&scope=${color.name.toLowerCase()}").response { request, response, result -> }
+            lights.update(color, false)
         }
         state = !state
         Thread.sleep(speed)
         run()
     }
+
 }
