+++ src/main/kotlin/tech/johari/gamelights/GameListener.kt
@@ -1,26 +1,27 @@
 package tech.johari.gamelights
 
-import com.github.kittinunf.fuel.Fuel
 import me.skreem.event.EventHandler
 import me.skreem.event.Listener
 import me.skreem.game.events.bomb.BombDefuseEvent
 import me.skreem.game.events.bomb.BombExplodeEvent
 import me.skreem.game.events.bomb.BombPlantEvent
 import me.skreem.game.events.round.RoundFreezetimeEvent
+import tech.johari.gamelights.lights.LightManager
+import tech.johari.gamelights.lights.LightSignal
 
 // Round playing -> nothing on
 // Planted -> Flash yellow [faster]
 // Defuse -> Green [slower]
 // Explodes -> All lights flashing [same pace as defuse]
-class Listener : Listener {
-    var currentAnimation : FlashThread? = null
+class GameListener(val lights: LightManager) : Listener {
 
+    var currentAnimation : FlashThread? = null
 
     private fun killAnimation() {
         if(currentAnimation != null) {
             currentAnimation!!.stop()
         }
-        Fuel.get("http://${Main.PI_LOCAL_IP}:${Main.PI_API_PORT}/ha-api?cmd=0&scope=all").response { request, response, result -> }
+        lights.update(LightSignal.ALL, false)
     }
 
     init {
@@ -36,7 +37,7 @@ class Listener : Listener {
     fun onBombPlantEvent(event: BombPlantEvent) {
         killAnimation()
 
-        currentAnimation = FlashThread(LightSignal.YELLOW, 500)
+        currentAnimation = FlashThread(lights, LightSignal.YELLOW, 500)
         currentAnimation!!.start()
     }
 
@@ -44,7 +45,7 @@ class Listener : Listener {
     fun onBombExplodeEvent(event: BombExplodeEvent) {
         killAnimation()
 
-        currentAnimation = FlashThread(LightSignal.ALL, 1000)
+        currentAnimation = FlashThread(lights, LightSignal.ALL, 1000)
         currentAnimation!!.start()
     }
 
@@ -52,7 +53,7 @@ class Listener : Listener {
     fun onBombDefuseEvent(event: BombDefuseEvent) {
         killAnimation()
 
-        currentAnimation = FlashThread(LightSignal.GREEN, 1000)
+        currentAnimation = FlashThread(lights, LightSignal.GREEN, 1000)
         currentAnimation!!.start()
     }
