+++ src/main/kotlin/tech/johari/gamelights/lights/LightSignal.kt
@@ -1,5 +1,7 @@
-package tech.johari.gamelights
+package tech.johari.gamelights.lights
 
 enum class LightSignal {
+
     RED, YELLOW, GREEN, ALL
+
 }
